# Conflicts:
#	src/main/java/com/jocoos/mybeautip/video/LegacyVideoService.java

+++ src/main/java/com/jocoos/mybeautip/client/aws/s3/AwsS3Handler.java
@@ -6,6 +6,7 @@ import com.jocoos.mybeautip.global.dto.FileDto;
 import com.jocoos.mybeautip.global.exception.BadRequestException;
 import com.jocoos.mybeautip.global.exception.S3UrlUploadException;
 import com.jocoos.mybeautip.support.RandomUtils;
+import io.micrometer.core.instrument.util.StringUtils;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Value;
@@ -15,7 +16,8 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.jocoos.mybeautip.global.util.FileUtil.getFilename;
+import static com.jocoos.mybeautip.global.constant.MybeautipConstant.TEST_FILE;
+import static com.jocoos.mybeautip.global.util.FileUtil.getFileName;
 import static com.jocoos.mybeautip.global.util.ImageUrlConvertUtil.getUri;
 
 @Component
@@ -60,7 +62,26 @@ public class AwsS3Handler {
             return "";
         }
 
-        String filename = getFilename(fileDto.getUrl());
+        String filename = getFileName(fileDto.getUrl());
+
+        String path = service.copy(
+                UrlDirectory.TEMP.getDirectory() + filename,
+                destination + filename);
+
+        return cloudFront + path;
+    }
+
+    public String copy(String fileUrl, String destination) {
+        if (StringUtils.isBlank(fileUrl)) {
+            return null;
+        }
+
+        String filename = getFileName(fileUrl);
+
+        // FIXME Test Code 에서 올라오는 File URL 은 실제 파일이 없기때문에 복사할 수가 없어서 회피용 코드 삽입. 뭔가 이 방법 말고 좋은 방법이 있다면... 고쳐주세요
+        if (filename.equals(TEST_FILE)) {
+            return cloudFront + destination + filename;
+        }
 
         String path = service.copy(
                 UrlDirectory.TEMP.getDirectory() + filename,
@@ -105,6 +126,14 @@ public class AwsS3Handler {
 
     public void delete(FileDto fileDto) {
         String filename = getUri(fileDto.getUrl());
+        service.delete(filename);
+    }
+
+    public void delete(String fileUrl) {
+        if (StringUtils.isBlank(fileUrl)) {
+            return;
+        }
+        String filename = getUri(fileUrl);
         service.delete(filename);
     }
 }
